refactor(RoleType): 패키지 변경
- 기존 constants 패키지 내부에 한 개의 클래스만 존재했기 때문에 model 클래스로 이동
